Small changes to dataset creation

+++ src/data/mapillary.py
@@ -92,7 +92,7 @@ def download_mapillary_image_information_by_bbox(bbox: Tuple[float], min_quality
     # sort_by=key enables pagination
     # todo: do we actually want to have pagination?
     url = (
-        'https://a.mapillary.com/v3/images?client_id={}&bbox={}&per_page=500&sort_by=key&min_quality_score={}' \
+        'https://a.mapillary.com/v3/images?client_id={}&bbox={}&per_page=5000&sort_by=key&min_quality_score={}'
     ).format(MAPILLARY_CLIENT_ID, bbox_str, min_quality_score)
 
     # download data from given URL
@@ -127,14 +127,17 @@ def download_mapillary_object_detection_by_key(image_key: str, download_dir: str
     Returns:
         None
     """
-    layer = "segmentations"
+    json_local_path = os.path.join(download_dir, "{}.json".format(image_key))
+    if not os.path.isfile(json_local_path):
+        layer = "segmentations"
+        # request for object detection layer of a certain image (given by image key)
+        url = (
+            "https://a.mapillary.com/v3/images/{}/object_detections/{}?client_id={}"
+        ).format(image_key, layer, MAPILLARY_CLIENT_ID)
+        r = requests.get(url, timeout=300)
+        data = r.json()
 
-    # request for object detection layer of a certain image (given by image key)
-    url = (
-        "https://a.mapillary.com/v3/images/{}/object_detections/{}?client_id={}" \
-    ).format(image_key, layer, MAPILLARY_CLIENT_ID)
-    r = requests.get(url, timeout=300)
-    data = r.json()
-
-    with open(os.path.join(download_dir, "{}.json".format(image_key)), 'w') as f:
-        json.dump(data, f)
+        with open(json_local_path, 'w') as f:
+            json.dump(data, f)
+    else:
+        print(json_local_path, "already exists. Skipping Download.")
